1. JPA 코드방식 변경 - 이전과 같지만 간략화됨

+++ src/main/java/com/study/sns/controller/PostController.java
@@ -69,7 +69,7 @@ public class PostController {
     }
 
     @GetMapping("/{postId}/likes")
-    public Response<Integer> likeCount(@PathVariable Integer postId,
+    public Response<Long> likeCount(@PathVariable Integer postId,
                                        Authentication authentication) {
 
         return Response.success(postService.likeCount(postId));
